+++ src/main/java/com/ttasjwi/cafe/controller/HomeController.java
@@ -1,13 +1,28 @@
 package com.ttasjwi.cafe.controller;
 
+import com.ttasjwi.cafe.domain.Article;
+import com.ttasjwi.cafe.repository.ArticleRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
 
 @Controller
 public class HomeController {
 
+    private final ArticleRepository articleRepository;
+
+    public HomeController(ArticleRepository articleRepository) {
+        this.articleRepository = articleRepository;
+    }
+
     @GetMapping
-    public String home() {
+    public String home(Model model) {
+        List<Article> articleList = articleRepository.findAll();
+        model.addAttribute("articles", articleList);
         return "home";
     }
 }